pyproject.toml
[project]
name = "smart-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
packages = [
    { include = "src" },
]
dependencies = [
    "gradio>=5.38.1",
    "pytube>=15.0.0",
    "pytubefix",
    "openai-whisper",
    "pymupdf",
    "sentence-transformers",
    "faiss-cpu",
    "dotenv",
    "langchain-openai",

]


#    "faiss-cpu>=1.11.0.post1",
#    "langchain>=0.3.26",
#    "openai>=1.97.1",
#    "pandas>=2.3.1",
#    "pymupdf>=1.26.3",
#    "fastapi>=0.100.0",
#    "langchain-community>=0.3.27",
#    "uvicorn>=0.35.0",
#    "python-multipart>=0.0.20",
#    "pypdf>=5.8.0",
#    "tiktoken>=0.9.0",
#    "langchain-openai>=0.3.28",
#    "langgraph>=0.5.4",
#    "python-docx>=1.2.0",
#    "chromadb>=1.0.15",
#    "pypdf2>=3.0.1",
#    "sentence-transformers>=5.0.0",
#    "numpy>=2.3.1",
#    "scikit-learn>=1.7.1",
#    "pydantic>=2.11.7",
#    "reportlab>=4.4.3",
#    "graphviz>=0.21",
#    "gradio>=5.38.1",
#    "langchain-huggingface>=0.3.1",
#    "langchain-cli>=0.0.36",
#    "isort>=6.0.1",
#    "unstructured>=0.18.11",